refactor: createCallbackManager
