+++ templates/types/streaming/express/src/controllers/chat.controller.ts
@@ -1,14 +1,9 @@
 import { Message, StreamData, streamToResponse } from "ai";
 import { Request, Response } from "express";
-import {
-  CallbackManager,
-  ChatMessage,
-  MessageContent,
-  Settings,
-} from "llamaindex";
+import { ChatMessage, MessageContent, Settings } from "llamaindex";
 import { createChatEngine } from "./engine/chat";
 import { LlamaIndexStream } from "./llamaindex-stream";
-import { appendEventData, appendToolData } from "./stream-helper";
+import { createCallbackManager } from "./stream-helper";
 
 const convertMessageContent = (
   textMessage: string,
@@ -52,32 +47,7 @@ export const chat = async (req: Request, res: Response) => {
     const vercelStreamData = new StreamData();
 
     // Setup callbacks
-    const callbackManager = new CallbackManager();
-    callbackManager.on("retrieve", (data) => {
-      const { nodes } = data.detail;
-      appendEventData(
-        vercelStreamData,
-        `Retrieving context for query: '${userMessage.content}'`,
-      );
-      appendEventData(
-        vercelStreamData,
-        `Retrieved ${nodes.length} sources to use as context for the query`,
-      );
-    });
-    callbackManager.on("llm-tool-call", (event) => {
-      const { name, input } = event.detail.payload.toolCall;
-      const inputString = Object.entries(input)
-        .map(([key, value]) => `${key}: ${value}`)
-        .join(", ");
-      appendEventData(
-        vercelStreamData,
-        `Using tool: '${name}' with inputs: '${inputString}'`,
-      );
-    });
-    callbackManager.on("llm-tool-result", (event) => {
-      const { toolCall, toolResult } = event.detail.payload;
-      appendToolData(vercelStreamData, toolCall, toolResult);
-    });
+    const callbackManager = createCallbackManager(vercelStreamData);
 
     // Calling LlamaIndex's ChatEngine to get a streamed response
     const response = await Settings.withCallbackManager(callbackManager, () => {
