+++ templates/types/streaming/nextjs/app/components/ui/chat/widgets/WeatherCard.tsx
@@ -1,4 +1,4 @@
-interface WeatherData {
+export interface WeatherData {
   latitute: number;
   longitude: number;
   generationtime_ms: number;
